+++ examples/fileio.py
@@ -4,8 +4,8 @@ Example script for file IO
 
 from pathlib import Path
 import numpy as np
-from pyomeca.types.markers import Markers3d
-from pyomeca.types.analogs import Analogs3d
+from pyomeca.obj.markers import Markers3d
+from pyomeca.obj.analogs import Analogs3d
 
 # Path to data
 DATA_FOLDER = Path('..') / 'tests' / 'data'
