+++ examples/rbdl.py
@@ -1,7 +1,7 @@
 import numpy as np
 
 from pyomeca.thirdparty import biorbd
-from pyomeca.types.generalized_coordinates import GeneralizedCoordinate
+from pyomeca.obj.generalized_coordinates import GeneralizedCoordinate
 from pyomeca.show.vtk import Model as PyoModel
 from pyomeca.show.vtk import Window as PyoWindow
